Feature srm 43185

See merge request operation-srm-watson/srm-mall-watsons!85

+++ src/main/java/org/srm/mall/other/api/dto/WatsonsAddressDTO.java
@@ -46,7 +46,7 @@ public class WatsonsAddressDTO extends AddressDTO {
     /**
      * 商城对应的地区等级
      */
-    private Integer mallRegionLevel;
+    private Long mallRegionLevel;
 
     /**
      * 百度地区编码
